Rename otf into gencert.
